S2 Week 2: Gmail-safe ErrorAlert HTML with error/audit/app log tails, fix FTS 7601 by adding SearchNormalized to full-text index, fix SubSectionSearchMatchRow int/long cast, add FTS availability guard in SubSectionService, configurable welcome-mail URL from SiteUrl, and move diagnostics middleware after authentication so Who fields are populated.

+++ NIGA.Centrum.Entity/DataModels/SubSectionSearchMatchRow.cs
@@ -2,7 +2,7 @@ namespace NIGA.Centrum.Entity.DataModels
 {
     public class SubSectionSearchMatchRow
     {
-        public long SubSectionId { get; set; }
+        public int SubSectionId { get; set; }
         public string SubSectionName { get; set; }
         public int? ParentSubSectionId { get; set; }
         public int Rank { get; set; }
